refactor: move time utils into renderer folder
