+++ src/renderer/containers/ReplayBrowser/ReplayFile.tsx
@@ -12,12 +12,12 @@ import TimerIcon from "@material-ui/icons/Timer";
 import { FileResult } from "@replays/types";
 import { stages as stageUtils } from "@slippi/slippi-js";
 import { colors } from "common/colors";
-import { convertFrameCountToDurationString, monthDayHourFormat } from "common/time";
 import _ from "lodash";
 import moment from "moment";
 import React from "react";
 
 import { DraggableFile } from "@/components/DraggableFile";
+import { convertFrameCountToDurationString, monthDayHourFormat } from "@/lib/time";
 import { getStageImage } from "@/lib/utils";
 
 import { TeamElements } from "./TeamElements";
