+++ src/replays/loadFile.ts
@@ -1,6 +1,7 @@
 import { GameStartType, MetadataType, SlippiGame } from "@slippi/slippi-js";
-import { fileToDateAndTime } from "common/time";
+import fs from "fs";
 import _ from "lodash";
+import moment from "moment";
 import path from "path";
 
 import { FileResult } from "./types";
@@ -40,4 +41,37 @@ export async function loadFile(fullPath: string): Promise<FileResult> {
   }
 
   return result;
+}
+
+function convertToDateAndTime(dateTimeString: moment.MomentInput): moment.Moment | null {
+  const asMoment = moment(dateTimeString);
+  if (asMoment.isValid()) {
+    return asMoment.local();
+  }
+
+  return null;
+}
+
+function fileToDateAndTime(
+  dateTimeString: string | undefined | null,
+  fileName: string,
+  fullPath: string,
+): moment.Moment | null {
+  const startAt = convertToDateAndTime(dateTimeString);
+  const getTimeFromFileName = () => filenameToDateAndTime(fileName);
+  const getTimeFromBirthTime = () => convertToDateAndTime(fs.statSync(fullPath).birthtime);
+
+  return startAt || getTimeFromFileName() || getTimeFromBirthTime() || null;
+}
+
+function filenameToDateAndTime(fileName: string): moment.Moment | null {
+  const timeReg = /\d{8}T\d{6}/g;
+  const filenameTime = fileName.match(timeReg);
+
+  if (filenameTime === null) {
+    return null;
+  }
+
+  const time = moment(filenameTime[0]).local();
+  return time;
 }
